remove NN stuff and make moves better

+++ Button.pde
@@ -1,11 +1,11 @@
 class Button {
   String label;
-  float x;    // top left corner x position
-  float y;    // top left corner y position
-  float w;    // width of button
-  float h;    // height of button
+  double x;    // top left corner x position
+  double y;    // top left corner y position
+  double w;    // width of button
+  double h;    // height of button
 
-  Button(String labelB, float xpos, float ypos, float widthB, float heightB) {
+  Button(String labelB, double xpos, double ypos, double widthB, double heightB) {
     label = labelB;
     x = xpos;
     y = ypos;
@@ -17,10 +17,10 @@ class Button {
     fill(218);
     stroke(141);
     rectMode(CORNER);
-    rect(x, y, w, h);
+    rect((float)x, (float)y, (float)w, (float)h);
     textAlign(CENTER, CENTER);
     fill(0);
-    text(label, x + (w / 2), y + (h / 2));
+    text(label, (float)(x + (w / 2)), (float)(y + (h / 2)));
   }
 
   boolean mouseIsOver() {
